requirements.txt
numpy==1.22.0
pandas==1.3.5
python-binance
wheel